+++ src/parser.tab.h
@@ -45,7 +45,7 @@
 extern int yydebug;
 #endif
 /* "%code requires" blocks.  */
-#line 599 "src/parser.y"
+#line 631 "src/parser.y"
 
 #include "ast.h"
 #include "parse_ctx.h"
@@ -171,7 +171,7 @@ typedef struct {
 #if ! defined YYSTYPE && ! defined YYSTYPE_IS_DECLARED
 union YYSTYPE
 {
-#line 624 "src/parser.y"
+#line 656 "src/parser.y"
 
     double number;
     char *text;